% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea-return-level.R
\name{alea_return_level}
\alias{alea_return_level}
\title{Return levels for ALEA fitted models}
\usage{
alea_return_level(object, return_period, ...)
}
\arguments{
\item{object}{An object returned by \code{alea_fit}.}

\item{return_period}{Numeric vector of return periods. Values must be finite
and greater than 1.}

\item{...}{Additional arguments reserved for future methods.}
}
\value{
A data frame with class \code{alea_return_level}.
}
\description{
Computes return levels for fitted ALEA models.
}
