% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice.R
\name{cast_dice}
\alias{cast_dice}
\title{Cast a dice}
\usage{
cast_dice(sides = 1:6, n_tries = 1, sum = FALSE)
}
\arguments{
\item{sides}{can either be an integer value (maximum value on a standard dice) or a vector of sides.}

\item{n_tries}{how often the dice should be cast.}

\item{sum}{whether the result of casting the dice should be summed.}
}
\description{
Cast a dice
}
\examples{
cast_dice()
cast_dice(5)
cast_dice(1:5) # same as example above
cast_dice(20, 3)
cast_dice(20, 3, TRUE)
}
