% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea-fit.R
\name{coef.alea_fit}
\alias{coef.alea_fit}
\title{Extract fitted model coefficients}
\usage{
\method{coef}{alea_fit}(object, type = c("standard", "internal"), ...)
}
\arguments{
\item{object}{An \code{alea_fit} object.}

\item{type}{Character scalar. Use \code{"standard"} to return user-facing
parameter names (\code{location}, \code{scale}, and \code{shape}). Use \code{"internal"} to
return distribution-specific internal parameter names.}

\item{...}{Further arguments ignored.}
}
\value{
A named numeric vector of fitted parameters.
}
\description{
Extract fitted model coefficients
}
