% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnd_dice.R
\name{d20}
\alias{d20}
\title{Cast a D20}
\usage{
d20(n_tries = 1, sum = FALSE)
}
\arguments{
\item{n_tries}{how often the dice should be cast.}

\item{sum}{whether the result of casting the dice should be summed.}
}
\description{
Cast a D20
}
