% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnd_dice.R
\name{d8}
\alias{d8}
\title{Cast a D8}
\usage{
d8(n_tries = 1, sum = FALSE)
}
\arguments{
\item{n_tries}{how often the dice should be cast.}

\item{sum}{whether the result of casting the dice should be summed.}
}
\description{
Cast a D8
}
