% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-diagnostics.R
\name{plot.alea_diagnostics}
\alias{plot.alea_diagnostics}
\title{Plot ALEA diagnostics}
\usage{
\method{plot}{alea_diagnostics}(x, type = c("status", "p_value"), ...)
}
\arguments{
\item{x}{An object of class \code{alea_diagnostics}.}

\item{type}{Character scalar. Plot type. One of \code{"status"} or \code{"p_value"}.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces diagnostic plots from an \code{alea_diagnostics} object.
}
\examples{
x <- c(42.1, 38.5, 51.3, 47.0, 62.4, 55.2, 49.8, 58.1,
       60.3, 45.9)
diagnostics <- alea_diagnostics(x)
plot(diagnostics, type = "status")

}
