% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-selection.R
\name{plot.alea_selection}
\alias{plot.alea_selection}
\title{Plot ALEA AI-assisted model selection}
\usage{
\method{plot}{alea_selection}(x, ...)
}
\arguments{
\item{x}{An object of class \code{alea_selection}.}

\item{...}{Additional arguments passed to methods. Currently unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces a probability ranking plot from an \code{alea_selection} object.
}
