% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea-fit.R
\name{print.alea_fit}
\alias{print.alea_fit}
\title{Print an ALEA fitted model}
\usage{
\method{print}{alea_fit}(x, ...)
}
\arguments{
\item{x}{An \code{alea_fit} object.}

\item{...}{Further arguments ignored.}
}
\value{
The input object, invisibly.
}
\description{
Print an ALEA fitted model
}
