% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_dice.R
\name{sicherman}
\alias{sicherman}
\title{Cast a Sicherman dice}
\usage{
sicherman(n_tries = 1, sum = FALSE)
}
\arguments{
\item{n_tries}{how often the dice should be cast.}

\item{sum}{whether the result of casting the dice should be summed.}
}
\description{
Cast a Sicherman dice
}
