% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea-fit.R
\name{coef.alea_fit}
\alias{coef.alea_fit}
\title{Extract fitted model coefficients}
\usage{
\method{coef}{alea_fit}(object, ...)
}
\arguments{
\item{object}{An \code{alea_fit} object.}

\item{...}{Further arguments ignored.}
}
\value{
A named numeric vector of fitted parameters.
}
\description{
Extract fitted model coefficients
}
