% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-gof.R
\name{plot.alea_gof}
\alias{plot.alea_gof}
\title{Plot ALEA goodness-of-fit results}
\usage{
\method{plot}{alea_gof}(x, type = c("statistic", "rank"), ...)
}
\arguments{
\item{x}{An object of class \code{alea_gof}.}

\item{type}{Character scalar. Plot type. One of \code{"statistic"} or \code{"rank"}.}

\item{...}{Additional arguments passed to methods. Currently unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Produces a goodness-of-fit plot from an \code{alea_gof} object.
}
